% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{load_ferrdb_genesets}
\alias{load_ferrdb_genesets}
\title{Load FerrDb genesets from an .rds file (recommended format)}
\usage{
load_ferrdb_genesets(rds_path)
}
\arguments{
\item{rds_path}{Path to FerrDbV3_RCD_genesets.rds.}
}
\value{
genesets list.
}
\description{
Load FerrDb genesets from an .rds file (recommended format)
}
