% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{rcd_rank_topk}
\alias{rcd_rank_topk}
\title{Rank top-k RCDs for each sample/cell using a chosen score matrix}
\usage{
rcd_rank_topk(score_mat, k = 3)
}
\arguments{
\item{score_mat}{RCD x samples matrix (e.g., res$net_tendency).}

\item{k}{Top-k.}
}
\value{
data.frame with sample, rank, RCD, score.
}
\description{
Rank top-k RCDs for each sample/cell using a chosen score matrix
}
