% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_one_set}
\alias{score_one_set}
\title{Score one gene set using activity matrix}
\usage{
score_one_set(A, genes, weights = NULL)
}
\arguments{
\item{A}{Activity matrix (genes x samples/cells).}

\item{genes}{Character vector of genes.}

\item{weights}{Named numeric vector of gene weights (optional).}
}
\value{
Numeric vector (length = ncol(A)).
}
\description{
Score one gene set using activity matrix
}
